requirements.txt
streamlit>=1.31.0
PyMuPDF>=1.23.0  # for fitz
google-generativeai>=0.3.0
umap-learn>=0.5.5
numpy>=1.24.0
scikit-learn>=1.3.0  # for NearestNeighbors
plotly>=5.18.0
python-dotenv>=1.0.0 